% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Act109.R
\name{retrieve_answer}
\alias{retrieve_answer}
\title{This function takes the user's selection and prints the selected response.}
\usage{
retrieve_answer(seleccion)
}
\arguments{
\item{seleccion}{user option (entero)}
}
\value{
Prints the selected answer.
}
\description{
This function takes the user's selection and prints the selected response.
}
\examples{
retrieve_answer(1)
retrieve_answer(5)
}
